add 90 degree isosceles triangle model
The crystal, reflector, etc. are contained in a .obj model. triangular.py just helps produce the SiPM and optical pad should you need those. There's also a FreeCAD project in the folder which can be exported to the .obj.

There is like an 8 Angstrom overlap in the simulation. Probably not a big deal but something to keep in mind.